Create CryptoCom.py
CryptoCom.py does the communicating with the exchange it defines methods to produce calls to the websocket url and rest endpoint url to operate the different types of actions needed for automated trading! it generates a signature in hmac sha256 to the exchange and makes a request from the class within the CryptoCom.py file!